check for cross compilation

+++ recipe/build.sh
@@ -4,9 +4,9 @@ cp $RECIPE_DIR/setup.py .
 
 export ARCTICDB_USING_CONDA=1
 
-rm $PREFIX/bin/protoc
-
-# do this only when cross compiling TODO
-ln -s $BUILD_PREFIX/bin/protoc $PREFIX/bin/protoc
+if [[ "$CONDA_BUILD_CROSS_COMPILATION" != "1" ]]; then
+  rm $PREFIX/bin/protoc
+  ln -s $BUILD_PREFIX/bin/protoc $PREFIX/bin/protoc
+fi
 
 CMAKE_BUILD_PARALLEL_LEVEL=1 $PYTHON -m pip install . -vv
